+++ packages/service/src/page/mypage/header/index.tsx
@@ -1,0 +1,13 @@
+import { MyPageFilter } from './filter';
+import { MyPageTabList } from './tab-list';
+
+import * as styles from './header.css';
+
+export function MyPageHeader() {
+  return (
+    <section className={styles.tabSection}>
+      <MyPageTabList />
+      <MyPageFilter />
+    </section>
+  );
+}
